+++ .eslintrc.json
@@ -1,0 +1,25 @@
+{
+    "root": true,
+    "extends": "airbnb-base/legacy",
+    "parserOptions": {
+        "ecmaVersion": 6
+    },
+    "globals": {
+        "empty": true,
+        "request": true,
+        "response": true,
+        "session": true,
+        "dw": true
+    },
+    "rules": {
+        "import/no-unresolved": "off",
+        "indent": ["error", 4, { "SwitchCase": 1, "VariableDeclarator": 1 }],
+        "func-names": "off",
+        "require-jsdoc": "error",
+        "valid-jsdoc": ["error", { "preferType": { "Boolean": "boolean", "Number": "number", "object": "Object", "String": "string" }, "requireReturn": false}],
+        "vars-on-top": "off",
+        "global-require": "off",
+        "no-shadow": ["error", { "allow": ["err", "callback"]}],
+        "max-len": "off"
+    }
+}
