+++ .stylelintignore
@@ -1,0 +1,4 @@
+cartridges/app_storefront_base/
+cartridges/bm_app_storefront_base/
+cartridges/modules/
+workspace/
