+++ cartridges/int_bloomreach_sfra/cartridge/scripts/services/blrServiceHelper.js
@@ -24,7 +24,8 @@ function mapUrlParams(data, url, prefix) {
         if (typeof data[key] === 'object') {
             params = mapUrlParams(data[key], params, localPrefix + key);
         } else {
-            params += '&' + localPrefix + key + '=' + data[key];
+            var value = (key === 'url' || key === 'ref_url') ? require('dw/crypto/Encoding').toURI(data[key]) : data[key];
+            params += '&' + localPrefix + key + '=' + value;
         }
     });
 
@@ -44,8 +45,8 @@ function performSearch(searchDetails) {
     var baseUrl = service.getURL();
     var accountId = libBloomreach.getPreference('AccountID');
     var domainKey = libBloomreach.getPreference('DomainKey');
-    var apiKey = libBloomreach.getPreference('ApiKey');
-    var path = baseUrl + '/api/v1/core/?account_id=' + accountId + '&auth_key=' + apiKey + '&domain_key=' + domainKey;
+    var authKey = libBloomreach.getPreference('AuthKey');
+    var path = baseUrl + '/api/v1/core/?account_id=' + accountId + '&auth_key=' + authKey + '&domain_key=' + domainKey;
     path = mapUrlParams(searchDetails, path);
     service.setURL(path);
 
@@ -65,8 +66,8 @@ function performSearchSuggestions(searchDetails) {
     var baseUrl = service.getURL();
     var accountId = libBloomreach.getPreference('AccountID');
     var domainKey = libBloomreach.getPreference('DomainKey');
-    var apiKey = libBloomreach.getPreference('ApiKey');
-    var path = baseUrl + '/api/v2/suggest/?account_id=' + accountId + '&auth_key=' + apiKey + '&catalog_views=' + domainKey;
+    var authKey = libBloomreach.getPreference('AuthKey');
+    var path = baseUrl + '/api/v2/suggest/?account_id=' + accountId + '&auth_key=' + authKey + '&catalog_views=' + domainKey;
     path = mapUrlParams(searchDetails, path);
     service.setURL(path);
 
@@ -81,25 +82,27 @@ function performSearchSuggestions(searchDetails) {
 function getRecommendations(recommendationParams) {
     var accountId = libBloomreach.getPreference('AccountID');
     var domainKey = libBloomreach.getPreference('DomainKey');
-    var apiKey = libBloomreach.getPreference('ApiKey');
+    var authKey = libBloomreach.getPreference('AuthKey');
+    var data = {};
 
     var service = serviceDefinition.init('bloomreach.http.widgets.api');
     var baseUrl = service.getURL();
     var path = baseUrl + recommendationParams.wty + '/' + recommendationParams.widget_id
         + '?account_id=' + accountId
-        + '&auth_key=' + apiKey
-        + '&domain_key=' + domainKey
-        + '&_br_uid_2=' + recommendationParams.brUid2;
+        + '&domain_key=' + domainKey;
 
     Object.keys(recommendationParams).forEach(function (key) {
-        if (['widget_id', 'wty', '_br_uid_2', 'brUid2'].indexOf(key) === -1) {
-            path += '&' + key + '=' + recommendationParams[key];
+        if (['widget_id', 'wty'].indexOf(key) === -1) {
+            data[key] = recommendationParams[key];
         }
     });
+
+    path = mapUrlParams(data, path);
 
     service.setURL(path);
     service.setRequestMethod('GET');
     service.addHeader('Content-Type', 'application/json');
+    service.addHeader('auth-key', authKey);
 
     return service.call();
 }
