+++ cartridges/int_bloomreach_sfra/cartridge/templates/default/pixels/pixelData.isml
@@ -1,17 +1,22 @@
 <iscontent type="text/html" charset="UTF-8" compact="true"/>
 <script>
 var br_data = br_data || {}; // psa
+
 <isloop items="${Object.keys(pdict.pixelData)}" var="key">
-    br_data.${key} = "${pdict.pixelData[key]}";
+    <isif condition="${key==='catalogs'}">
+        br_data.${key} = <isprint value="${pdict.pixelData[key]}" encoding="off" />;
+    <iselse>
+        br_data.${key} = "${pdict.pixelData[key]}";
+    </isif>
 </isloop>
 
 (function() {
-var brtrk = document.createElement('script');
-brtrk.type = 'text/javascript';
-brtrk.async = true;
-brtrk.src = "//cdn.brcdn.com/v1/br-trk-${pdict.pixelData.acct_id}.js";
+    var brtrk = document.createElement('script');
+    brtrk.type = 'text/javascript';
+    brtrk.async = true;
+    brtrk.src = "//cdn.brcdn.com/v1/br-trk-${pdict.pixelData.acct_id}.js";
 
-var s = document.getElementsByTagName('script')[0];
-s.parentNode.insertBefore(brtrk, s);
+    var s = document.getElementsByTagName('script')[0];
+    s.parentNode.insertBefore(brtrk, s);
 })();
 </script>
