+++ package.json
@@ -1,0 +1,82 @@
+{
+  "name": "int_bloomreach_sfra",
+  "version": "23.1.0",
+  "description": "Bloomreach link cartridge providing support search services",
+  "main": "index.js",
+  "engines": {
+    "node": ">=4.0"
+  },
+  "scripts": {
+    "test": "sgmf-scripts --test test/unit/**/*.js",
+    "test:integration": "sgmf-scripts --integration 'test/integration/**/*.js'",
+    "compile:scss": "sgmf-scripts --compile css",
+    "compile:js": "sgmf-scripts --compile js",
+    "compile:fonts": "node bin/Makefile compileFonts",
+    "build": "npm run compile:js && npm run compile:scss",
+    "lint": "npm run lint:css && npm run lint:js",
+    "lint:css": "sgmf-scripts --lint css",
+    "lint:js": "sgmf-scripts --lint js",
+    "uploadCartridge": "sgmf-scripts int_bloomreach && sgmf-scripts --uploadCartridge int_bloomreach_sfra"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/SalesforceCommerceCloud/link_bloomreach.git"
+  },
+  "author": "Astound Commerce",
+  "license": "MIT",
+  "homepage": "https://astoundcommerce/link_bloomreach.git#readme",
+  "devDependencies": {
+    "@babel/core": "^7.6.3",
+    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
+    "@babel/preset-env": "^7.6.3",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^8.0.6",
+    "babel-preset-env": "^1.7.0",
+    "chai": "^4.2.0",
+    "chai-subset": "^1.6.0",
+    "css-loader": "^0.28.11",
+    "debug": "^4.1.1",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb-base": "^5.0.3",
+    "eslint-plugin-import": "^1.16.0",
+    "eslint-plugin-sitegenesis": "~1.0.0",
+    "isml-linter": "^5.26.4",
+    "mocha": "^5.2.0",
+    "mocha-junit-reporter": "^1.23.1",
+    "node-sass": "^4.12.0",
+    "postcss-loader": "^2.1.6",
+    "popper.js": "^1.14.7",
+    "properties-parser": "^0.3.1",
+    "proxyquire": "1.7.4",
+    "sass-loader": "^7.3.1",
+    "sgmf-scripts": "^2.3.0",
+    "shelljs": "^0.8.3",
+    "should": "^13.2.3",
+    "sinon": "^1.17.7",
+    "style-loader": "^0.21.0",
+    "stylelint": "^8.4.0",
+    "stylelint-config-standard": "^17.0.0",
+    "stylelint-scss": "^2.5.0",
+    "url": "^0.11.0",
+    "utf-8-validate": "^5.0.2"
+  },
+  "dependencies": {
+    "bootstrap": "4.3.1",
+    "cleave.js": "^1.5.3",
+    "flag-icon-css": "^2.9.0",
+    "font-awesome": "^4.7.0",
+    "jquery": "^3.5.1"
+  },
+  "browserslist": [
+    "last 2 versions",
+    "ie >= 10"
+  ],
+  "babel": {
+    "presets": [
+      "env"
+    ]
+  },
+  "paths": {
+    "base": "../storefront-reference-architecture/cartridges/app_storefront_base/"
+  }
+}
